+++ src/modules/sbstudio/plugin/model/mixins.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Generic, Sequence, TypeVar, cast
+from typing import TYPE_CHECKING, Generic, Sequence, TypeVar
 
 if TYPE_CHECKING:
     from bpy.types import bpy_prop_collection_idprop  # exists only for type checking
@@ -37,9 +37,7 @@ class ListMixin(Generic[T]):
             next_entry = self.entries[index + 1]
 
             if self._on_active_entry_moving_down(this_entry, next_entry):
-                cast("bpy_prop_collection_idprop[T]", self.entries).move(
-                    index, index + 1
-                )
+                self.entries.move(index, index + 1)
                 self.active_entry_index = index + 1
 
     def move_active_entry_up(self) -> None:
@@ -52,9 +50,7 @@ class ListMixin(Generic[T]):
             this_entry = self.entries[index]
 
             if self._on_active_entry_moving_up(this_entry, prev_entry):
-                cast("bpy_prop_collection_idprop[T]", self.entries).move(
-                    index, index - 1
-                )
+                self.entries.move(index, index - 1)
                 self.active_entry_index = index - 1
 
     def _on_active_entry_moving_down(self, this_entry, next_entry) -> bool: